backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.7.2
motor==3.6.0
pymongo==4.9.1
pydantic==2.10.3
python-dotenv==1.0.1
